+++ src/main/java/net/hyper/weaponry/datagen/ModItemTagProvider.java
@@ -3,8 +3,8 @@ package net.hyper.weaponry.datagen;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
 import net.fabricmc.fabric.api.tag.convention.v2.ConventionalItemTags;
-import net.hyper.weaponry.item.ModItems;
-import net.hyper.weaponry.util.ModTags;
+import net.hyper.weaponry.registry.ModItems;
+import net.hyper.weaponry.registry.ModTags;
 import net.minecraft.item.Items;
 import net.minecraft.registry.RegistryWrapper;
 import net.minecraft.registry.tag.ItemTags;
@@ -44,19 +44,41 @@ public class ModItemTagProvider extends FabricTagProvider.ItemTagProvider {
         getOrCreateTagBuilder(ModTags.Items.GREATSWORDS)
                 .add(ModItems.WOODEN_GREATSWORD)
                 .add(ModItems.STONE_GREATSWORD)
+                .add(ModItems.COPPER_GREATSWORD)
                 .add(ModItems.IRON_GREATSWORD)
                 .add(ModItems.GOLDEN_GREATSWORD)
                 .add(ModItems.DIAMOND_GREATSWORD)
                 .add(ModItems.NETHERITE_GREATSWORD);
+        getOrCreateTagBuilder(ItemTags.SWORDS)
+                .add(ModItems.COPPER_SWORD);
+        getOrCreateTagBuilder(ItemTags.PICKAXES)
+                .add(ModItems.COPPER_PICKAXE);
+        getOrCreateTagBuilder(ItemTags.SHOVELS)
+                .add(ModItems.COPPER_SHOVEL);
+        getOrCreateTagBuilder(ItemTags.AXES)
+                .add(ModItems.COPPER_AXE);
+        getOrCreateTagBuilder(ItemTags.HOES)
+                .add(ModItems.COPPER_HOE);
+
+        getOrCreateTagBuilder(ConventionalItemTags.NUGGETS)
+                .add(ModItems.COPPER_NUGGET);
 
         getOrCreateTagBuilder(ItemTags.WEAPON_ENCHANTABLE)
                 .addTag(ModTags.Items.GREATSWORDS);
         getOrCreateTagBuilder(ItemTags.BREAKS_DECORATED_POTS)
                 .addTag(ModTags.Items.GREATSWORDS);
         getOrCreateTagBuilder(ItemTags.DURABILITY_ENCHANTABLE)
-                .addTag(ModTags.Items.GREATSWORDS);
+                .addTag(ModTags.Items.GREATSWORDS)
+                .add(ModItems.COPPER_HELMET)
+                .add(ModItems.COPPER_CHESTPLATE)
+                .add(ModItems.COPPER_LEGGINGS)
+                .add(ModItems.COPPER_BOOTS);
         getOrCreateTagBuilder(ItemTags.VANISHING_ENCHANTABLE)
-                .addTag(ModTags.Items.GREATSWORDS);
+                .addTag(ModTags.Items.GREATSWORDS)
+                .add(ModItems.COPPER_HELMET)
+                .add(ModItems.COPPER_CHESTPLATE)
+                .add(ModItems.COPPER_LEGGINGS)
+                .add(ModItems.COPPER_BOOTS);
         getOrCreateTagBuilder(ItemTags.SHARP_WEAPON_ENCHANTABLE)
                 .addTag(ModTags.Items.GREATSWORDS);
         getOrCreateTagBuilder(ConventionalItemTags.TOOLS)
@@ -64,10 +86,34 @@ public class ModItemTagProvider extends FabricTagProvider.ItemTagProvider {
         getOrCreateTagBuilder(ConventionalItemTags.MELEE_WEAPON_TOOLS)
                 .addTag(ModTags.Items.GREATSWORDS);
         getOrCreateTagBuilder(ConventionalItemTags.ENCHANTABLES)
-                .addTag(ModTags.Items.GREATSWORDS);
+                .addTag(ModTags.Items.GREATSWORDS)
+                .add(ModItems.COPPER_HELMET)
+                .add(ModItems.COPPER_CHESTPLATE)
+                .add(ModItems.COPPER_LEGGINGS)
+                .add(ModItems.COPPER_BOOTS);
         getOrCreateTagBuilder(ItemTags.SWORD_ENCHANTABLE)
                 .addTag(ModTags.Items.GREATSWORDS);
         getOrCreateTagBuilder(ItemTags.FIRE_ASPECT_ENCHANTABLE)
                 .addTag(ModTags.Items.GREATSWORDS);
+        getOrCreateTagBuilder(ItemTags.TRIMMABLE_ARMOR)
+                .add(ModItems.COPPER_HELMET)
+                .add(ModItems.COPPER_CHESTPLATE)
+                .add(ModItems.COPPER_LEGGINGS)
+                .add(ModItems.COPPER_BOOTS);
+        getOrCreateTagBuilder(ItemTags.HEAD_ARMOR)
+                .add(ModItems.COPPER_HELMET)
+                .add(ModItems.COPPER_CHESTPLATE)
+                .add(ModItems.COPPER_LEGGINGS)
+                .add(ModItems.COPPER_BOOTS);
+        getOrCreateTagBuilder(ConventionalItemTags.ARMORS)
+                .add(ModItems.COPPER_HELMET)
+                .add(ModItems.COPPER_CHESTPLATE)
+                .add(ModItems.COPPER_LEGGINGS)
+                .add(ModItems.COPPER_BOOTS);
+        getOrCreateTagBuilder(ItemTags.EQUIPPABLE_ENCHANTABLE)
+                .add(ModItems.COPPER_HELMET)
+                .add(ModItems.COPPER_CHESTPLATE)
+                .add(ModItems.COPPER_LEGGINGS)
+                .add(ModItems.COPPER_BOOTS);
     }
 }