+++ src/main/java/net/hyper/weaponry/registry/ModArmorMaterials.java
@@ -1,0 +1,42 @@
+package net.hyper.weaponry.registry;
+
+import net.hyper.weaponry.Weaponry;
+import net.minecraft.item.ArmorItem;
+import net.minecraft.item.ArmorMaterial;
+import net.minecraft.item.Items;
+import net.minecraft.recipe.Ingredient;
+import net.minecraft.registry.Registries;
+import net.minecraft.registry.Registry;
+import net.minecraft.registry.entry.RegistryEntry;
+import net.minecraft.sound.SoundEvents;
+import net.minecraft.util.Identifier;
+import net.minecraft.util.Util;
+
+import java.util.EnumMap;
+import java.util.List;
+import java.util.function.Supplier;
+
+public class ModArmorMaterials {
+    public static final RegistryEntry<ArmorMaterial> COPPER = registerArmorMaterial("copper",
+            () -> new ArmorMaterial(Util.make(new EnumMap<>(ArmorItem.Type.class), map -> {
+                map.put(ArmorItem.Type.BOOTS, 1);
+                map.put(ArmorItem.Type.LEGGINGS, 2);
+                map.put(ArmorItem.Type.CHESTPLATE, 3);
+                map.put(ArmorItem.Type.HELMET, 4);
+                map.put(ArmorItem.Type.BODY, 4);
+            }), 8, ModSounds.ITEM_ARMOR_EQUIP_COPPER, () -> Ingredient.ofItems(Items.COPPER_INGOT),
+                    List.of(new ArmorMaterial.Layer(Identifier.of(Weaponry.MOD_ID, "copper"))), 0,0));
+    public static final RegistryEntry<ArmorMaterial> NETHERITE = registerArmorMaterial("netherite",
+            () -> new ArmorMaterial(Util.make(new EnumMap<>(ArmorItem.Type.class), map -> {
+                map.put(ArmorItem.Type.BOOTS, 3);
+                map.put(ArmorItem.Type.LEGGINGS, 6);
+                map.put(ArmorItem.Type.CHESTPLATE, 8);
+                map.put(ArmorItem.Type.HELMET, 3);
+                map.put(ArmorItem.Type.BODY, 19);
+            }), 8, SoundEvents.ITEM_ARMOR_EQUIP_NETHERITE, () -> Ingredient.ofItems(Items.NETHERITE_INGOT),
+                    List.of(new ArmorMaterial.Layer(Identifier.of(Weaponry.MOD_ID, "netherite"))), 3.0f,0.1f));
+
+    public static RegistryEntry<ArmorMaterial> registerArmorMaterial(String name, Supplier<ArmorMaterial> material) {
+        return Registry.registerReference(Registries.ARMOR_MATERIAL, Identifier.of(Weaponry.MOD_ID, name), material.get());
+    }
+}