+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2a+1.21 Changelog:
+- backported copper gear
+- added copper greatswords
+- added copper and netherite horse armors
+- fixed iron and golden greatswords not being able to be smelted down into nuggets 
+- fixed wooden greatswords not being usable as fuel 
+- some internal code changes
+
 ## 0.1c+1.21 Changelog:
 - moved hand positioning and visual disabling of the offhand to its own tag
 - added a whitelist tag to keep specific items visual in the offhand while holding greatswords